Moved unused functions to EverythingElse
