+++ layers/EverythingElse/packages/tags/src/TypeAPIs/TagTypeAPIInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Tags\TypeAPIs;
+
+use PoPSchema\Taxonomies\TypeAPIs\TaxonomyTypeAPIInterface;
+
+/**
+ * Methods to interact with the Type, to be implemented by the underlying CMS
+ */
+interface TagTypeAPIInterface extends TaxonomyTypeAPIInterface
+{
+    public function getTagBase(): string;
+}
