+++ layers/Schema/packages/tags/src/TypeAPIs/TagTypeAPIInterface.php
@@ -30,6 +30,5 @@ interface TagTypeAPIInterface extends TaxonomyTypeAPIInterface
     public function getCustomPostTags(string | int $customPostID, array $query = [], array $options = []): array;
     public function getCustomPostTagCount(string | int $customPostID, array $query = [], array $options = []): int;
 
-    public function getTagBase(): string;
     public function setPostTags(string | int $customPostID, array $tags, bool $append = false);
 }
